// src/features/resume/types/interfaces.ts
export interface IExperience {
    position: string;
    startingMonth: string;
    startingYear: string;
    endingMonth: string;
    endingYear: string;
    companyName: string;
    summary: string;
    isCurrentEmployment: boolean;
    location: string
}

export interface IEducation {
    school: string;
    timeline: string;
    location: string;
    summary: string;
}